Initial commit of UTF-8 string processing library
This is part of a series of small libraries written in C that do one
thing. This one is for processing UTF-8 text. It is partially complete
with the core functionality mostly implemented and the project structure
completely implemented (all of these small C libraries have the same
overall structure).